fix/removed check constraint from role.py and moved nw endpoints to v2 folder
