+++ app/api/v2/api.py
@@ -7,7 +7,7 @@ from app.api.v2.routers.health import router as health_router
 from app.api.v2.routers.webhooks import router as webhooks_router
 from app.api.v2.routers.callback_scheduler import agents_router as cb_agents_router
 from app.api.v2.routers.callback_scheduler import calls_router as cb_calls_router
-from app.api.api_v1.endpoints.workspace import v2_router as workspace_v2_router
+from app.api.v2.routers.workspace import v2_router as workspace_router
 from app.api.v2.routers.hipaa import flows_router as hipaa_flows_router
 from app.api.v2.routers.hipaa import workspace_router as hipaa_workspace_router
 
@@ -19,6 +19,6 @@ v2_router.include_router(batch_calls_router)
 v2_router.include_router(webhooks_router)
 v2_router.include_router(cb_agents_router)
 v2_router.include_router(cb_calls_router)
-v2_router.include_router(workspace_v2_router, prefix="/workspace", tags=["Workspace Settings"])
+v2_router.include_router(workspace_router, prefix="/workspace", tags=["Workspace Settings"])
 v2_router.include_router(hipaa_flows_router)
 v2_router.include_router(hipaa_workspace_router)
